requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose[cryptography]
python-multipart
bcrypt
pydantic[email]
python-dotenv